pyproject.toml
[project]
name = "storyboard"
version = "0.1.0"
description = "Storyboard scene generation CLI tool"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.3.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.scripts]
storyboard = "storyboard.cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["storyboard"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
